add types, improve comment, move into remote connection resolver into distinct file
