+++ modules/server/src/network/index.ts
@@ -9,6 +9,7 @@ import {
 import urljoin from 'url-join';
 import { NetworkAggregationError } from './errors';
 import { fetchGql } from './gql';
+import { createRemoteConnectionResolvers } from './resolvers/connections';
 import { createNetworkAggregationTypeDefs } from './schema';
 import { NetworkAggregationConfig, NetworkAggregationConfigInput } from './types';
 
@@ -115,57 +116,6 @@ const fetchRemoteSchemas = async ({
 		});
 
 	return schemaResults;
-};
-
-/**
- *
- */
-const createRemoteConnectionResolvers = (networkConfigs: NetworkAggregationConfig[]) => {
-const connectionStatus = {
-	OK: 'OK',
-	ERROR: 'ERROR',
-};
-const checkRemoteConnectionStatus = async (url: string) => {
-	/**
-	 * recommended way to health check gql server is to run the `__typename` query that every server has
-	 * very small query with no additional params to, so using GET is not a concern for the GQL server
-	 * adds recommended pre flight header to make sure Apollo doesn't block request by CSRF protection
-	 */
-	const healthCheckQuery = '?query=%7B__typename%7D';
-	const fullUrl = urljoin(url, healthCheckQuery);
-
-	try {
-		const pong = await axios.get(fullUrl, {
-			headers: { 'apollo-require-preflight': 'true', 'Content-Type': 'application/json' },
-		});
-
-		// only need to check the expected query contacted the server and returns successfully (just a health check)
-		if (pong.data?.__typename) {
-			return connectionStatus.OK;
-		} else {
-			throw Error('no data object returned from GQL server __typname query');
-		}
-	} catch (error) {
-		console.error(error);
-		return connectionStatus.ERROR;
-	}
-};
-const createRemoteConnectionResolvers = async (networkConfigs: NetworkAggregationConfig[]) => {
-	/**
-	 * Promise.all is safe because we handle errors in checkRemoteConnectionStatus
-	 */
-	const remoteConnections = await Promise.all(
-		networkConfigs.map(async (config) => {
-			const { schema, ...configProperties } = config;
-			// includes default inbuilt GQL server types
-			const availableAggregations = schema ? schema.getTypeMap() : [];
-			// connection status
-			const status = await checkRemoteConnectionStatus(configProperties.graphqlUrl);
-			return { ...configProperties, availableAggregations, status };
-		}),
-	);
-
-	return remoteConnections;
 };
 
 /**
